Refactor Checkbox component and cell stratification group filter states (#37)
* In progress migration of filter state and checkbox components to own dirs

* Finish typing and refactoring Checkbox and filter states

* format

* undo accidental change to dna acc constant

* format
