+++ src/components/dna-accessibility/DnaAccessibility.tsx
@@ -14,10 +14,10 @@ import TitleBar from '../../TitleBar';
 
 import {
   higlassDnaAccessState,
-  sampleSelectionState,
   variantYScaleState,
   dnaAccessXDomainWithAssembly,
 } from '../../state';
+import { sampleSelectionState } from '../../state/filter-state';
 import {
   dnaAccessLabelStyleState,
   dnaAccessRowNormState,
