+++ src/components/enhancer-genes/EnhancerGenesPlot.js
@@ -17,7 +17,7 @@ import { useChromInfo } from '../../ChromInfoProvider';
 import { useShowTooltip } from '../../TooltipProvider';
 import { getCategories, getSamples } from './enhancer-gene-plot-helper';
 
-import { sampleSelectionState } from '../../state';
+import { sampleSelectionState } from '../../state/filter-state';
 import {
   focusRegionAbsWithAssembly,
   focusRegionRelState,
